==> Research/6.1.Sentiment_2020.ipynb <==
#!pip install pandas numpy emoji langdetect transformers torch


import pandas as pd
import torch
from transformers import AutoTokenizer, AutoModelForSequenceClassification, Trainer, TrainingArguments
import csv
import os
from transformers import AutoModel, AutoTokenizer
import torch
from sklearn.cluster import KMeans
import numpy as np
import torch
import torch.nn.functional as F
import urllib.request
from scipy.special import softmax

df = pd.read_csv('2020_Classified_Balanced.csv')
df.head()

## **Roberta Sentiment MODEL**
https://huggingface.co/cardiffnlp/twitter-roberta-base-sentiment

# load model and tokenizer
task = 'sentiment' # set the task to sentiment
MODEL = f"cardiffnlp/twitter-roberta-base-{task}"# specify the model name based on the task
tokenizer = AutoTokenizer.from_pretrained(MODEL, use_fast=False)
model = AutoModelForSequenceClassification.from_pretrained(MODEL, use_safetensors=True).to('cuda')


device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print("Using device:", device)

# load label mapping
mapping_link = f"https://raw.githubusercontent.com/cardiffnlp/tweeteval/main/datasets/{task}/mapping.txt"
with urllib.request.urlopen(mapping_link) as f:
    html = f.read().decode('utf-8').split("\n")
    csvreader = csv.reader(html, delimiter='\t')
labels = [row[1] for row in csvreader if len(row) > 1]
print("Labels:", labels)
# (Barbieri et al., 2020; cardiffnlp, 2020)

# preprocess text (username and link replacement)
# to follow consistency with the model's training data
def preprocess(text): # (Barbieri et al., 2020; cardiffnlp, 2020)
    new_text = []
    for t in str(text).lower().split(" "):
        t = '@user' if t.startswith('@') and len(t) > 1 else t
        t = 'http' if t.startswith('http') else t
        new_text.append(t)
    return " ".join(new_text)

# function to get sentiment
def getSentiment(text):
    text = preprocess(text)
    encoded_input = tokenizer(text, return_tensors='pt', truncation=True, padding=True, max_length=128).to('cuda')
    with torch.no_grad():
        output = model(**encoded_input)
        scores = output.logits[0].detach().cpu().numpy()
        scores = softmax(scores)
    pred_id = np.argmax(scores)
    # return label and scores
    return labels[pred_id], scores
# (Barbieri et al., 2020; cardiffnlp, 2020)

# apply the function to the DataFrame
sentOutput = df["tweet"].apply(getSentiment)
# split tuple results into separate columns
df["sentiment_label"] = sentOutput.apply(lambda x: x[0])
df["sentiment_probs"] = sentOutput.apply(lambda x: x[1])
df.head()
# (Barbieri et al., 2020; cardiffnlp, 2020)

labels = ['negative', 'neutral', 'positive']
sentiment_counts = df.groupby(['label', 'sentiment_label']).size().unstack(fill_value=0)
# convert counts to proportions for each political leaning
sentiment_avg = sentiment_counts.div(sentiment_counts.sum(axis=1), axis=0)

# Save results
df.to_csv("2020_sentiment_robertaBase.csv", index=False)

**Results**

# average sentiment probabilities by political leaning
print("\nAverage Sentiment Distribution by Political Leaning:\n")
sentiment_avg
