==> Research/8.1.Toxicity_2020.ipynb <==
#!pip install pandas numpy emoji langdetect transformers torch



import pandas as pd
from transformers import AutoTokenizer, AutoModelForSequenceClassification, AutoConfig
import torch
import numpy as np
from scipy.special import softmax
import urllib.request, csv

df = pd.read_csv('classified samples/2020_Classified_Balanced.csv')
df.info()

# preprocess text 
# to follow consistency with the model's training data
def preprocess(text): 
    new_text = []
    for t in str(text).lower().split(" "):
        t = '@user' if t.startswith('@') and len(t) > 1 else t
        t = 'http' if t.startswith('http') else t
        new_text.append(t)
    return " ".join(new_text)

# Load model and tokenizer
MODEL = 'textdetox/twitter-xlmr-toxicity-classifier' 
tokenizer = AutoTokenizer.from_pretrained(MODEL)
model = AutoModelForSequenceClassification.from_pretrained(MODEL, use_safetensors=True).to('cuda')
labels = ["non-toxic", "toxic"] #(Hanu, 2020; Lai-Lopez et al., 2025)

# function to get hate speech label and probabilities
def get_toxicity_label(text):
    text = preprocess(text)
    encoded_input = tokenizer(
          text, 
          return_tensors="pt", 
          truncation=True, 
          padding=True, 
          max_length=128
         ).to('cuda')
    with torch.no_grad():
     output = model(**encoded_input)
     scores = output.logits[0].detach().cpu().numpy()
     probs = softmax(scores)
    
    pred_id = np.argmax(probs)
    return labels[pred_id], probs

# Apply the function to the DataFrame
results = df["tweet"].apply(get_toxicity_label)
# split tuple results into separate columns
# Split results
df["toxicity_label"] = results.apply(lambda x: x[0])
df["toxicity_probs"] = results.apply(lambda x: x[1])
df.head()

df_probs = pd.DataFrame(df["toxicity_probs"].tolist(), columns=labels)
df = pd.concat([df, df_probs], axis=1)
#(Hanu, 2020; Lai-Lopez et al., 2025)

**Results**

if "label" in df.columns:
    avg_tox = df.groupby("label")[labels].mean().reset_index()
   # Save output if needed
df.to_csv("analysis/2020_toxicity_results.csv", index=False)
#(Hanu, 2020; Lai-Lopez et al., 2025)

print("\n Average Toxicity Probabilities by Political Leaning:")
print(avg_tox)
